Moved AbstractDAOTest to dpc-testing
