+++ dpc-api/src/test/java/gov/cms/dpc/api/jdbi/IpAddressDAOTest.java
@@ -1,7 +1,7 @@
 package gov.cms.dpc.api.jdbi;
 
 import gov.cms.dpc.api.entities.IpAddressEntity;
-import gov.cms.dpc.common.hibernate.AbstractDAOTest;
+import gov.cms.dpc.testing.AbstractDAOTest;
 import gov.cms.dpc.common.hibernate.auth.DPCAuthManagedSessionFactory;
 import io.hypersistence.utils.hibernate.type.basic.Inet;
 import org.junit.jupiter.api.BeforeEach;
